:sparkles: Add camera position management

+++ Core/source/lualibrary.cpp
@@ -124,8 +124,8 @@ namespace LuaLibrary {
   int lua_draw_sprite(lua_State* L) {
     unsigned long long sprite_id = static_cast<unsigned long long>(lua_tonumber(L, -4));
     int subimage = static_cast<int>(lua_tonumber(L, -3));
-    float x = static_cast<float>(lua_tonumber(L, -2));
-    float y = static_cast<float>(lua_tonumber(L, -1));
+    double x = lua_tonumber(L, -2);
+    double y = lua_tonumber(L, -1);
 
     lua_pushstring(L, "sprite_database");
     lua_gettable(L, LUA_REGISTRYINDEX);
@@ -136,7 +136,7 @@ namespace LuaLibrary {
     WindowManager* window_manager = static_cast<WindowManager*>(lua_touserdata(L, -1));
 
     SpriteAsset* spr = sprite_database->get_sprite_by_id(sprite_id);
-    window_manager->draw(spr, subimage, static_cast<int>(x), static_cast<int>(y));
+    window_manager->draw(spr, subimage, x, y);
     return 0;
   }
